Pub Sub/pyproject.toml
[tool.poetry]
name = "pubsub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.34.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
google-cloud-pubsub = "^2.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
